+++ tests/unit/services/courthouseService.test.ts
@@ -1,3 +1,4 @@
+export {};
 const courthouseService = require('../../../server/src/services/courthouseService');
 const courthouseRepository = require('../../../server/src/repositories/courthouseRepository');